python/requirements.txt
Pillow
pyrealsense2
numpy
matplotlib
mlxtend
dlib
opencv-python